+++ src/App/Router.php
@@ -45,11 +45,11 @@ class Router
 		$this->routeCollector->addRoute(['GET'],         '/host-meta',           Module\HostMeta::class);
 		$this->routeCollector->addRoute(['GET'],         '/hostxrd',             Module\HostMeta::class);
 		$this->routeCollector->addRoute(['GET'],         '/nodeinfo/1.0',        Module\NodeInfo::class);
-		$this->routeCollector->addRoute(['GET'],         '/xrd',                 Module\WebFinger::class);
+		$this->routeCollector->addRoute(['GET'],         '/xrd',                 Module\Xrd::class);
 		$this->routeCollector->addGroup('/.well-known', function (RouteCollector $collector) {
 			$collector->addRoute(['GET'], '/host-meta'       , Module\WellKnown\HostMeta::class);
 			$collector->addRoute(['GET'], '/nodeinfo[/1.0]'  , Module\NodeInfo::class);
-			$collector->addRoute(['GET'], '/webfinger'       , Module\WebFinger::class);
+			$collector->addRoute(['GET'], '/webfinger'       , Module\WellKnown\WebFinger::class);
 			$collector->addRoute(['GET'], '/x-social-relay'  , Module\WellKnown\XSocialRelay::class);
 		});
 	}
