+++ src/Module/WellKnown/XSocialRelay.php
@@ -13,8 +13,6 @@ class XSocialRelay extends BaseModule
 {
 	public static function rawContent()
 	{
-		parent::rawContent();
-
 		$app = self::getApp();
 		$config = $app->getConfig();
